[AD-262] Add additional tests for CASE
